[Easy] Only run one build per PR (#65)
We are currently building twice for each pull request (once because it's a PR and once because of the branch matching rule).

This PR adds a line to the travis file that prevents it from building on every branch. This should reduce the build time per PR in half.

Thanks @luarx for pointing this out.