[Driver] Defining an interface for the driver<->solver interaction (#66)
This PR defines a rust interface that the order_processor will interact with to invoke the solving engine. The solving engine can be implemented in different ways

Mocked (for unit tests)
Simple Solver (#58)
Tom's sophisticated solve (#59)
Let's discuss the interface and data structures here. I included the mock implementation so that we can start using this interface.

Addresses #57