Making sure our models don't know about price finding internals (#119)
I didn't like that our models "know" (as in import a dependency) about how the price finding decides to serialize account balances when it passes input data to the python solver (by exposing the ought to be private accountId and tokenId methods). This is an implementation detail to the linear optimization solver that should not be leaked to the basic data model.

I also didn't like that State had to know about Solution (in order to update balances it took a reference to the Solution object). While we could argue that Solution is a basic data model that should be moved into the models submodule (instead of being in price_finding as it is right now), there is no need for the two to be coupled. update_balances can be a free method on the order_processor, which already has to know about both Solution and State.

This PR moves the code back and cleans up the imports.

In order to allow simpler unit testing (e.g. not having to create 1k * 30 balances for each state) I made the num_tokens an ivar on the State object. This allows more concise testing.

Test Plan:
```
cargo test
e2e test
```