[Driver] Adding a nonce to the state hash (#120)
Currently two state roots with the same balances will have the same hash. This can introduce problems as in our data-model we are only filtering by hash in our query and not by the state_index.

One way to solve this is to introduce a nonce into the hash (this way the state_index is part of the entropy of the hash).
Alternatively we could also query both the currentStateRoot as well as the current stateIndex from the smart contract and filter for a combination of state_index and state_hash. However this would require another contract call each time we query for balance state.

One disadvantage of the proposed approach is that the hashes in our e2e tests now depend on the state_index that they'd been applied on. This means that changing the order of the tests or adding another test, might have consequences to all later e2e test hashes.

There are other ways to solve this issue if it becomes problematic (e.g. computing the expected hash as part of the test).