[Listener] processing StandingSellOrderBatch events (#132)
This PR contains logic for processing StandingSellOrderBatch events. It depends on gnosis/dex-contracts#97 and assumes that the order details are emitted as blobs.

For now it relies on unit tests to ensure functionality is correct.

Test Plan
Once depending PR is landed, update submodule and run placeStandingOrder script. Observe that standing orders are added to the database.