[E2E Test] wait up to 30s for ganache to open port (#204)
Seeing a lot of e2e tests fail because ganache is not yet started (after 1s sleep). 

E.g. https://travis-ci.org/gnosis/dex-services/jobs/565485559

This was introduced in #196 as we got rid of the retry library that was used before. This PR introduces are tool that allows waiting for a port to open (with timeout in ms).

### Test Plan

See travis pass and be no longer flaky.