GraphNode reader (#213)
At the moment, the only way to interact with the low level postgres representation from The Graph is via the [Store trait](https://github.com/graphprotocol/graph-node/blob/master/graph/src/components/store.rs#L650).

This trait is very elaborative in that it not only handles both read and write operations, it also covers both chain related storage as well as entity related storage.

The only thing we are interested on the driver side (and long term also on the listener side) is to read Entities. This means, we don't want to setup the entire dependency graph that the main graph component requires (GQL adapter, ethereum adapter, etc), but instead want to be able to instantiate a reader component just from a postgres connection (with a read-only user).

This PR takes the large store component from the graph repo and removes all write and chain related parts, leaving only a simple interface to query entities (code is copied). This interface will be used to implement database access inside the driver.

Long term, it would be nice if we could push such a separation of concerns inside the graph's main repository as their monolithic storage design seems quite inflexible.


### TestPlan

Make sure we can instantiate a version of this reader in the driver's main function just using a postgres connection.