176 Part 2 - Delete all python (#227)
In summary: 

- Most of the changes here are deleting the entire `event_listener` directory (i.e. the directory containing all the python files).

- 

This is the final phase of removing the django-event-listener. 

Removing all mongoDB queries from the end to end tests. When reviewing, I would think it is important to pay special attention to the remaining E2E-tests. That is, we should ensure that our end to end tests are still covering all the cases we had hoped to. Ideally, if the individual Graph Listener PRs had covered the tests they were supposed to, we should have to worry about this, but it would not hurt (since this is the final phase of transition), to keep this in mind while reviewing!

Test Plan - e2e-testing also do a quick global search on the project for "python" and "mongo"

closes #176