Refactoring account balances to be represented as a sparse HashMap (#235)
We currently represent balances inside the AccountState model as a continuous array with values in the form of 

```
acc1_token1, acc1_token2, ... acc1_tokenN, acc2_token1 ..., accM_tokenN
```

While this works well for a small number of accounts, this doesn't scale to the new dfusion model, where our address space is 256 bits. 

The number of participants in a batch is upper bound by the number of orders in a batch. We only need to pass the balances to the solver of accounts that are participating in the current auction. This suggests that we should be using a sparse HashMap rather than a list over the full address space to represent balances.

For backwards compatibility, I kept the interface that allows creating the AccountState object from a continuous list and also kept the way we store the state in the GraphDb (this will likely change as we plan to increase or address space to at least 24 bits there as well).

The new driver component will be able to expose another new method, passing in the sparse HashMap directly.


### Testplan

existing unit tests and e2e tests