Only passing the balances that are required by the solver to python (#237)
At the moment, we serialise the entire account state for the solver's input json, where really we only need the balances of the accounts that are trading.

This PR changes it to make it more sparse. Note, that we need to provide the buyBalance as well since otherwise, I'm running into the following python error:

```sh
[INFO: D_verify_SNARK.py:60 | checkAccountBalances()]  === ACCOUNT BALANCES ===
| Traceback (most recent call last):
|   File "./batchauctions/scripts/optimize_e2e.py", line 125, in <module>
|     args.solverTimelimit)
|   File "./batchauctions/scripts/optimize_e2e.py", line 93, in run
|     return verifySNARK(jsonFile_03, 1e13)
|   File "/app/batchauctions/scripts/verification/D_verify_SNARK.py", line 192, in run
|     c3 = checkAccountBalances(accounts, orders)
|   File "/app/batchauctions/scripts/verification/D_verify_SNARK.py", line 71, in checkAccountBalances
|     _acc[o['accountID']][o['buyToken']] += Decimal(o['execBuyAmount'])
| KeyError: 'token1'
```

cc @twalth3r I guess the feature of implicitly assuming 0 was added after April 16th (which is the version I'm currently pegged to)? Anyway, no big deal - we can do this logic in rust.

### Test Plan

Same as #236