[PoC] Get (relevant) Auction Data from Smart Contract (#249)
The driver's smart contract interface should implement the function get_auction_data which takes auction ID (or batch index) and returns the state of accounts along with all the relevant orders in the requested auction.

This implementation invokes the StableX contract's "getAuctionElementsFunction" which returns every order that has ever been placed in a byte encoded format. Thus, much of the work here involves deserializing the byte-encoded orders and filtering out those which are irrelevant for the current batch.