Submit Solution With Computed Objective Value 2 (#300)
This PR implements retrieving the objective value from the stableX contract before submitting the solution. This allows the driver to verify the solution before committing any gas to submit a solution with its objective value to the contract. The contract can then skip verification if the reported objective value is lower than the latest winning solution and save gas in case of multiple solutions racing to be submitted.

Because of recent changes to the smart contract, changes to the naive solver were required. As an interim solution, the naive solver for Snapp and StableX were split into two implementation. They intend to be merged at a later date.

This PR addresses issue #279.

This PR was originally two separate PRs #296 and #297