Add Snapp Objective Value Calculation (#303)
Since the solver no longer calculates executed utility. Unfortunately this is required for computing the solution objective value which needs to be submitted with a solution for Snapp.

This PR is a first step to addressing this by calculating the Snapp objective value (the total executed utility of the orders) for submitting solutions.

This might help with #302 and with removing the objective_value field from the solution. This will also be nice for merging the two naive solvers together so it no longer needs to calculate this.

### Test Plan

Unit tests.

### Commit history

* added order executed utility calculation

* added snapp objective value computation

* submit solution with snapp objective value for snapp

* cargo fmt