[Easy] Remove Solution Objective Value Field (#305)
Remove the solution objective value field as it is no longer used. We now calculate the Snapp objective value for submitting solutions to Snapp and we use the contract to evaluate the objective value in StableX.

This affects PR #302 and might cause merge conflicts. Depends on #304.

### Test Plan

unit and e2e tests

### Commit History

* removed objective_value from solution

* fixed warnings from rebase

* cargo fmt and clippy