[CI] Add rinkeby e2e test to CI (#312)
This is a somewhat bold proposal which should be carefully discussed. My goal is to have an acceptance test that will only make travis green if the PR that is about to be merged will work with the currently deployed rinkeby version of the contracts (using Tom's solver and a simple trade example).

To do this I'm running the stablex-rinkeby image and am placing two limit orders (+deposits) on the deployed rinkeby contract. We then wait for `secondsRemainingInBatch` + 30 seconds so that the batch can close and the solver can propose a solution.

We then check that there are no errors in the logs.

Here are some problems with the current approach
1. There is no way to fast-forward rinkeby so in the worst case we will sleep 5 minutes.
2. The travis solver might be time racing with the solver that is already deployed on AWS making this test not very deterministic (however we'd still test that everything until submitSolution would be working e2e)
3. Just checking "there are no errors" might be too weak, but other checks might not work due to non determinism (e.g. other orders in the same batch might influence the result)
4. Orders are done with real rinkeby addresses, which would have to be funded constantly (1 OWL, DAI per commit). Maybe this could be circumvented by also requesting withdraws (however that would add to the time it takes to know whether the test passes).

If time is the only concern we could also create a job that only runs on master and tags a given commit as "stable" if it passes the long running test (so that our PRs can be merged quicker). We could then set up alerts if "stable" gets stuck and the frontend could depend on "stable".

I'd still much prefer this over a develop branch which we manually merge in fixed intervals as I'd like to have the frontend team work with and thus test our latest versions (otherwise no one would and so the merge from develop to master or release-candidate would be arbitrary and not guarantee any quality). 
But of course we want to make sure that our commits are not obviously broken.