Moving travis file to multi-staged build (#354)
Goal of this PR is to avoid building our rust project on each travis step, and instead have a single build stage that creates a minified image only containing the binary.

Each subsequent E2E test would then download that image and not have to do any compilation.

Results:

||Total Time|Run Time|Link|
|---|---|---|---| 
|Before|57min|21min|https://travis-ci.org/gnosis/dex-services/builds/624108072
|After|42min|23min|https://travis-ci.org/gnosis/dex-services/builds/624341622

While the run time (the time we have to wait to see the green checkmark) didn't go down, we are using 15 minutes less total resource for each run (reducing the time we block other projects).

In order to reduce this time even further (I had a run with 16 minutes run time, 35 minutes total time) we need to reuse the locally built binaries and copy them into the docker image (instead of building them inside docker). For that to work all images need to have a xenial base, so it was starting to make this PR very large.

I will take a step back and create a xenial based solver image inside the dex-solver repo and then cycle back to this repo to simplify the Dockerfiles we have for our rust image (we should just need one that is based on the solver).