Using dex-solver image and having a single conditional base image (#362)
This PR builds on top of the previous work and dramatically cleans up the amount of build scripts we have in our repo. We now basically have two orthogonal choices
1. Which solver image to use (default a vanilla ubuntu without the dex-solver)
2. Which rust bas image to use (one that compiles everything or one that uses local binaries)

With this we can support both local development as well as cached travis builds and releases that will be deployed on AWS.

### Test Plan
I updated the README and made sure I can still run the system locally with both naive and linear solver. The rest is up to Travis.