Fix travis cache (#388)
I noticed that our rust build are no longer cached.

The rust cache was defined in the global section on top of the `travis.yml` (above the individual sections for each build stage). 
This worked fine as long as the rust stages (build/unit tests) did not define their own cache directories themselves (it overrides instead of merges the list from the global section).

This PR moves the cached directories into the respective section.

### Test Plan

Expecting to see shorter build times as of the second build.