Do not retry solving the instance if  failed (#421)
Closes #417 

The solver is running deterministically. Therefore, there is no point in calling it again with the same problem. Thus if it failed to produce a result, we may as well consider this batch as unsolvable.

I'd like to do the same for smart contract rejections. However, right now we can still have transient failures (e.g. 503 error from the full node) and thus it is not safe to rely on it. 

With an upgrade to ethcontracts we might be able to check for revert errors (and we should retry transient failures).

### Test Plan
Added a unit test.