Use serde to (de)serialize the solver input and output json
closes #284

This is less error prone and less repetitive than managing the json by
hand.

We not change which fields are handled.

We remove code handling both strings and numbers in parse_price_value
because these fields are always strings.

We remove some assertions of a specific error descriptions in favor of
only asserting the error kind. Error descriptions are human readable.
The code should not rely on specific strings.