Replacing mock-it with mockall (#471)
If I remember correctly, when we first evaluated the available mocking frameworks in rust (using [this](https://asomers.github.io/mock_shootout/) list), there was no convenient one available for stable rust. We therefore chose [mock-it](https://github.com/nathanielsimard/mock-it), which was on of the few libraries for stable.

One drawback with mock-it is that it requires writing our own mock-implementations of all the traits we are depending on. This creates a lot of repetitive boilerplate code.

A recent re-evaluation of the landscape made me discover [mockall](https://github.com/asomers/mockall) which seems to be able to auto derive the mock implementations using macros and checks almost all green boxes  on the comparison matrix.

This PR replaces mock-it with mockall, kill a bunch of boilerplate code and rewrites the tests in the new framework.

### Test Plan

`cargo tests`