[Fix Orderbook filter] No longer requiring the ODERBOOK_FILTER env (#483)
The PR #476 is currently throwing errors in production.

Due to the fact, that we do not set the env: ORDERBOOK_FILTER, the code runs the fallback and sets ORDERBOOK_FILTER=""; Due to this, we are receiving the following error:
ERRO [stablex] Error parsing orderbook filter: EOF while parsing a value at line 1 column 0

This PR changes the fallback for the env from "" to "{}". Now the parsing is fine and we don't get any error logs.

We could have also set ORDERBOOK_FILTER in production, but I would prefer this fallback solution to work.