433 batched orders (#484)
Currently, when reading the order book we are retrieving all orders ever placed in a single call to the smart contract. This can fail if there are too many orders even with high gas limits.
This commit changes the reading method to read orders in batches of configurable size. This is already support by the smart contract.