Fix panic
The naive solver only finds one match between two orders. For this match it sets the prices in a hashmap but not for the others. The second for loop which is responsible for handling the fee then iterates over all orders, not just the two matched orders which causes a panic when the prices hashmap is accessed for tokens that weren't part of the match.