[Easy] store each solver result in a different folder (#541)
This PR makes it so that we don't override the result file in every auction. Instead we store them in separate folders similar to how we do it for input files.

This will allow us to sync the results with S3 and visualize them on the public S3 bucket.

### Test Plan

Run local instance with mainnet solver for two batches. See both result files on disk.