Do not set block number when reading order book
We are currently having an issue where the node does not support queries
with block numbers in the past because it is pruning the state.
Until we decide whether this is fixable on the node's side we decided to
remove the fixed block number. This can in theory lead to reading the
orderbook in an inconsisent state when two pages come from different
batches. If this happens then we already cannot solve the auction
because there is no time left seeing as the batch changed.
A consequence of this is that we previously had a panic in the page
reading code when we notice that despite fixing the block number there
is still an inconsistency. The panic has been turned into a warning.
We could instead keep the check and return an error instad of panicking
but this makes the code more convoluted and there can still be
inconsistencies we cannot detect like including orders that no longer
exist.

related to #544