Initial Kraken API implementation (#562)
This is the initial implementation of a Kraken HTTP API to be used by the price oracle. The full integration with the solver will come piecemeal

Note that the API is currently not used anywhere, but its a way to build up to a larger feature in manageable reviewable PRs.

### Test Plan

CI with new unit tests. Also, a test that does not run by default is included and can be run by:
```
cargo test online_kraken_api -- --ignored --nocapture
```

### Commit History

* Initial Kraken API implementation

* fixed typo

* addressing comments