Inlcude batchID in instance/result file name (#620)
This PR makes it so that our instance/result files include the batchID and are also split up by date. We now create instance filenames in the following format:

`instances/{day}/instance_{batchId}_{datetime}.json`

and results in the following format:

`results/{day}/instance_{batchId}_{datetime}/<json|visualization>`

### Test Plan

Tested locally that files are created and read accordingly. For synchronisation with S3 I have to slightly adjust the sync script to not be too strict about the matching.