Open solver integration part1 (#680)
This PR integrates the open-solver.

While building the docker images, there will be choice between building on the closed-source standard-solver or the open-source standard solver. Depending on the choice different images will be created: either the `stablex-binary-public` or `stablex-binary-private`. These new images, we can then deploy with kubernetes, as a second PR will show.

Please note that this process of creating two driver dockers takes much more time and hence Travis is significantly slower. But the parallelization of these builds is out of the scope of this PR.

testplan: travis only