[Easy] New solver config: standard solver should also use external prices (#700)
In the solver-team, we realized that the solutions found under consideration of the external prices are almost always better than the other solutions.

Hence, we decided to use external price feeds always. (cp: https://gnosisinc.slack.com/archives/CS6NV17HP/p1585220425012100)

The main difference between the standard and the fallback solver will be the whitelist of tokens, which I will configure in gnosis-staging.