Use BatchExchangeViewer contract for order fetching (#723)
This PR changes the contract that we use to fetch the on-chain orderbook from BatchExchange to BatchExchangeViewer. We made a few gas optimizations inside the viewer contract, which should allow for a larger page size and thus fast orderbook fetching.

We needed to update the dex-contracts submodule to get the latest deployments of the BatchExchangeViewer and need to wait on a hotfix in ethcontracts which will allow functions with void return types in the ABI.

### Test Plan

Run driver locally and see orderbook fetching on 4G in ~30s.