[getOpenOrderbook 1] Split PaginatedAuctionDataReader (#742)
We are using the `PaginatedAuctionDataReader` abstraction to process pages that we fetch via `getEncodedOrdersPaginated`. I would like to reuse that abstraction for the open orderbook fetching which filters valid orders on-chain.

The pagination logic for this way of fetching is done differently (also on-chain). This PR therefore splits the functionality of PaginatedAuctionDataReader into two parts

1) AuctionDataReader which just applies pages and knows nothing about pagination
2) PaginatedAuctionDataReader which uses 1) for page application and on top of that keeps track of page info

This way we will be able to reuse component 1) in the next PR

I slightly changed the semantics of the component in so far as that instead of returning the elements inside the unfiltered page (which would not be available in the open orderbook fetching) we know make the pagination field optional. None will indicate that there is no next page.

### Test Plan

Adjusted unit tests and made sure they are still passing.