[getOpenOrderbook 2] Use onchain filter to fetch orders (#748)
Draft to get some early feedback. Still missing unit tests and proper wiring (not sure if I should make it configurable of if we want to get rid of the existing Orderbook reader, which is probably safer).

This PR introduces a new Reader implementation that uses the BatchExchangeViewer and in particular its methods that filter order on0chain (instead of on the client). The main benefit is that this methods allows to query balance for a future batch (e.g. the one that starts in the next block) and will thus allow us to fetch the orderbook from the last block of the previous batch rather than on the first block of the batch we are actually trying to solve

### Test Plan

Will add a unit test to the AuctionDataReader. Also running this locally and making sure that at least number of orders/users/tokens are correct. Would probably be cool to already use this in combination with the Diff tool and see that it runs smooth locally.