Retrieve orderbook on last block of solving batch. (#736)
This PR introduces code to search for the last block of the solving batch and retrieve the orderbook on that batch. This makes so the open solver and EVM-scheduler solvers that start later than system time solvers don't potentially retrieve an invalid orderbook that contains solution trades.

### Test Plan

CI and let it run in staging for a bit to see what happens. We have tried this before, but ran into issues where we were not able to query state at a specific block. The theory is that the load balancer was connecting different requests to different nodes that had not yet seen the block.

If we run into this, we can try enforcing a keep-alive policy to ensure that these queries get done to the same node.