Do not use a stream for event based orderbook
Using a error stream prevents use from using load balanced nodes because
if the node changes under us the new node will not have the log filter
that was set up by the stream.
Instead we always query all past logs of the last X blocks, delete the
previous events from this range and then apply the one.
This works even when the node switches and makes it easier to implement
retries on errors in the futures.