[Easy] Don't log batches that ran out of time as successful (#784)
https://github.com/gnosis/dex-services/pull/656#discussion_r395529456 identified and explained the issue. 

We are currently already seeing a bunch of auctions in which we run out of time fetching the orderbook (cf. [kibana](https://logs.gnosisdev.com/goto/75c991c4167005acff9e14a8b72db20a)).

However this is not at all reflected in our top-level grafana metrics or alerts. Reason is that such batches are still recorded as being successfully processed.

This PR makes it so that we don't log a successful processing event in this case and adds some more docs to make the expectation of each event stage more explicit for future maintainers.

### Test Plan

Once merged, expect to see more "downtime" in the staging Grafana dashboard.