[Streamed Orderbook] Fix batchId off-by-one error (#786)
This is a ~pretty~ extremely hacky short-term fix for #785 with the goal to keep on getting more consistency results and elminate known issues.

To summarize, `orderbook_for_batch` currently returns the balances that are valid in `batch - 1` and the orders that are valid in `batch`. This PR changes it to also return the orders that are valid in `batch-1`. Then, we need to increment the batchID we get passed in from `OrderbookReading` as this is already the index of the batch we are trying to solve (and for which we want the orderbook).

I also renamed the variable in the OrderbookReading trait to make it more explicit which batch this ID is for.

I'm happy to make/accept a cleaner fix if you guys have an idea.

### Test Plan

1. The first bug was that we would get errors if an event is received in the first 30s of a batch (since this would advance last_batch_id and then result assertion errors that we are querying on a past batch id). I tested locally that this is no longer the case (https://rinkeby.etherscan.io/tx/0x7556fe7b1ddfcb51d1b72cbc67c3060bad9dc5cae3639cd540f74c363e8c4942)
2. When creating an order and a deposit in the same batch the order would get correctly included by the streamed orderbook, but the deposit would be ignored. I tested locally that this is no longer the case (tx from above + deposit in the same batch https://rinkeby.etherscan.io/tx/0x63f94ce7cd43aaf2295da66315fffc2dffd43612f1308ec5dd3725717bf6707c)