Use private solver instead of naive solver on travis
Now that we have a public, open solver it is no longer important to test
that the naive solver works. This commit changes the ci test from the
naive solver to the private solver.
We might want to remove the naive solver code in the future or replace
it with something from pricegraph.