Use 256 bit integer to represent balance instead of 128
In the smart contract account balances are serialized as 256 bit
integers but our current implementation uses 128 bit integers instead
and asserts that balances are never larger than the maximum 128 bit
integer.
It is more correct to use the same number type which allows us to
represent all balances and gets rid of the panic.