Use arbitrary size integers for event based orderbook
Currently we use 256 bit integers, like the smart contract to keep track
of account balances and how they change with trades.
This commit changes this to use arbitrary size integers. This has the
advantage of making the code simpler because we can get rid of some
error conditions and no longer need to keep track of positive and
negative balance changes separately.
This is also a disadvantage because the new code is less accurate: It
does not detect some invalid event chains that the old would have
detected.
Another disadvantage is that arbitrary sized integers by design allocate
and live on the heap which makes the new code slower. Although I do not
expect the change in speed to matter.

This commit also updates the test for solution reversion to get more
coverage.