Improve event based orderbook tests with macros
This makes the tests more readable by making the event definitions
shorter and resembling human language more. We still show exactly the
same information as before when the events are created just in a
different format.

In another commit I am going to try turning the tests into a more
declarative style which would get ride of passing `state` to the macro.