[Easy] Reduce solver logging output (#938)
We ran into issues with the logging service that the solver was outputting a lot which was being grouped into a single extremely large log document and caused issues.

This PR addresses this by reducing the logging level for the solver as suggested here https://github.com/gnosis/dex-solver/issues/257

### Test Plan

CI (to make sure the option is accepted), next time we get a solver error ensure the log message is not so large. I'm open to suggestions on how to improve this test plan.