Refactor Order to include all order information
Previously we had split up the order information into
StableXAuctionElement which contained sell_token_balance, valid_from,
valid_until and Order which contained the modified numerator.
Now Order contains all order specific information: valid_from,
valid_until and the original numerator, denominator and remaining
amount.

This is a cleaner approach as we propagate the original order
information for as long as possible. Only when the order is passed to
the solver we calculate the new numerator and denominator so that there
is no remaining amount.

While working on this I noticed that Order derived a couple of unused
traits which I have removed.