Add price estimator scaffolding (#959)
This is the initial commit for the rust price estimator. It sets up the
smart contract but does not yet create an orderbook or handle the http
requests we want.

For the http server libraries I have looked at:
* https://github.com/http-rs/tide
* https://github.com/hyperium/hyper
* https://github.com/seanmonstar/warp
* https://github.com/tomaka/rouille

tide is based on async std. I did not choose it because the readme says
that is not ready for production.
rouille is the http server we use for metrics in the driver. It does not
use async which makes it slower than the async frameworks (see benchmark
in their readme). This speed difference is probably not important for
our task but we might as well use an async framework from the get go.
warp is based on hyper. It easier to use because hyper is too low
level. Both are linked in the tokio readme.