Set up orderbook
Add an Orderbook to the price estimator and set up one task to update
and another task to serve http requests.
To handle requests the get_reduced_orderbook method on Orderbook is
goign to be used and the resulting pricegraph::Orderbook will be moved
into the function that handles the request.