Refactor TransitiveOrder to include price and capacity
The transitive orderbook as returned from price estimator is defined
over price and volume. Pricegrpah has this internally but converts it
into buy and sell amounts instead.
By expressing this in terms of methods we can pick the internal
representation as we like while making use of the struct more
convenient.