Fix pricegraph edge cases (#1018)
`reduce_overlapping_transitive_orderbook` was missing a check for the
token pair validity which would cause a panic in bellman_ford::search
when the quote token didn't exist.

`fill_transitive_orders` had an infinite loop with an empty orderbook
and the token pair buying and selling the fee token. `path::find_path`
would in this case keep returning `Some`.