Revert "fix transitive orderbook ask/bid direction (#1029)"
This reverts commit f9abf3ea0b4c31724874d63a04de5a85c0ebeed8.

The reason for this is that when I tested the rust price estimator
against the node price estimator again just now I noticed that the asks
and bids are reverted: visualizing the orderbook for pair A-B in the
node price estimator is the same as pair B-A in the rust price
estimator.
The reason we made the PR I am reverting here is given in https://github.com/gnosis/dex-services/pull/1021#discussion_r441311440 .

I am not yet sure where the bug is but until we find it we should
revert.