Calculate rounding buffer like solver
Currently we are applying a rounding buffer as a fixed factor but this
is not how the solver works and thus ends up usually too conservative. By
calculating the rounding buffer like the solver does we can get a more
realistic result.
This commit adds the rounding buffer calculation but does not yet apply
it or actually retrieve prices.