Merge transitive orders of same price
Having several orders at the same price caused ui issues in the
frontend. To fix this we merge orders of the same price by adding their
volumes.
Conceptually it is not clear whether merged or unmerged is more
correct as merging loses the information that the same price was reached
through different paths through the order graph (i think).
For this reason I implemented this in price-estimator and not
pricegraph.