Remove price from TokenBaseInfo (#1075)
Step 4 of plan outlined in #1068

This PR introduces a new type `TokenOverrideInfo` that is  read from the config `TOKEN_DATA` for overriding token information, such as symbol or price. `TokenBaseInfo` can then only contain `symbol` and `decimals` as prices are now fetched exclusively via the `PriceSource` trait.

Originally I planned to make all fields optional, however this will create some weirdness when we override symbol but still need to go on-chain for fetching decimals (the current abstraction doesn't work as nicely and would involve introducing a lot more optionals and `if let Some(x)` logic). It seems reasonable to override symbol and decimals in any case when doing a price override (they never change and symbol will serve as documentation which token we are overriding).

### Test Plan

CI